+++ R/winsorize.R
@@ -3,7 +3,9 @@
 #' Winsorization of data.
 #'
 #' @param data Dataframe or vector.
-#' @param threshold The amount of Winsorization.
+#' @param threshold The amount of winsorization.
+#' @param verbose Toggle warnings.
+#' @param ... Currently not used.
 #'
 #' @examples
 #' library(correlation)
@@ -11,13 +13,13 @@
 #' winsorize(iris$Sepal.Length, threshold = 0.2)
 #' winsorize(iris, threshold = 0.2)
 #' @export
-winsorize <- function(data, threshold = 0.2) {
+winsorize <- function(data, ...) {
   UseMethod("winsorize")
 }
 
 
 #' @export
-winsorize.factor <- function(data, threshold = 0.2) {
+winsorize.factor <- function(data, threshold = 0.2, ...) {
   data
 }
 
@@ -28,13 +30,20 @@ winsorize.character <- winsorize.factor
 winsorize.logical <- winsorize.factor
 
 #' @export
-winsorize.data.frame <- function(data, threshold = 0.2) {
-  sapply(data, winsorize, threshold = threshold)
+winsorize.data.frame <- function(data, threshold = 0.2, verbose = TRUE, ...) {
+  sapply(data, winsorize, threshold = threshold, verbose = verbose)
 }
 
-
+#' @rdname winsorize
 #' @export
-winsorize.numeric <- function(data, threshold = 0.2) {
+winsorize.numeric <- function(data, threshold = 0.2, verbose = TRUE, ...) {
+  if (threshold < 0 || threshold > 1) {
+    if (isTRUE(verbose)) {
+      warning("'threshold' for winsorization must be a scalar between 0 and 1. Did not winsorize data.", call. = FALSE)
+    }
+    return(data)
+  }
+
   # TODO: This function should go in the future data package.
   y <- sort(data)
   n <- length(data)